Coerces bioacousticsquote objects to strings suitable for LaTeX

This function coerces bioacousticsquote objects to strings suitable for rendering in LaTeX. Quotes and (potential LaTeX) sources are placed within suitable "epigraph" output format via the sprintf function.

Usage

as.latex(quotes, form = "\\epigraph{%s}{%s}\n\n", cite = TRUE)

Arguments

quotes

an object of class bioacousticsquote returned from functions such as search_quotes or bioacousticsquote

form

structure of the LaTeX output for the text (first argument) and source (second argument) passed to sprintf

cite

logical; should the cite field be included in the source output?

Value

character vector of formatted LaTeX quotes

See also

Author

Phil Chalmers

Examples

ll <- search_quotes("Hempton")
as.latex(ll)
#> [1] "\\epigraph{Seuketat is the Eskimo word for ear-of-the-animal. This is what we must become if we want to truly listen.}{Gordon Hempton, Earth is a Solar Powered Jukebox}\n\n"                                                      
#> [2] "\\epigraph{How can we listen for something that we have not yet heard? We can't. When we listen in this way, selectively, we are actually practicing controlled impairment.}{Gordon Hempton, Earth is a Solar Powered Jukebox}\n\n"