Display a randomly chosen statistical quote.

Display a randomly chosen statistical quote.

Usage

bioacousticsquote(
  ind = NULL,
  pattern = NULL,
  tag = NULL,
  source = NULL,
  topic = NULL
)

# S3 method for class 'bioacousticsquote'
print(x, cite = TRUE, width = NULL, ...)

# S3 method for class 'bioacousticsquote'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

ind

Integer or character. If 'ind' is missing, a random quote is chosen from all quotations. If 'ind' is specified and is an integer, return the ind^th quote. If 'ind' is specified and is character, use it as the 'pattern'.

pattern

Character string. Quotes are subset to to those which match the pattern in the quote text.

tag

Character string. Quotes are subset to those matching the specified tag.

source

Character string. Quotes are subset to those matching the specified source (person).

topic

Deprecated. Use 'tag' instead. Only kept for backward compatibility.

x

object of class 'bioacousticsquote'

cite

logical; should the cite field be printed?

width

Optional print width parameter

...

Other optional arguments, unused here

row.names

see as.data.frame

optional

see as.data.frame

Value

A character vector containing one quote. It is of class bioacousticsquote for which an S3 print method will be invoked, and for which other methods are available.

See also

Examples

set.seed(1234)
bioacousticsquote()
#> 
#> There was never a King like Solomon, Not since the world began; But
#> Solomon talked to a butterfly, As a man would talk to a man.
#> --- Rudyard Kipling, The Butterfly that Stamped
bioacousticsquote(10)
#> 
#> A great silence is spreading over the natural world even as the sound
#> of man is becoming deafening.
#> --- Bernie Krause, The Great Animal Orchestra
bioacousticsquote("Seuketat")
#> 
#> Seuketat is the Eskimo word for ear-of-the-animal. This is what we must
#> become if we want to truly listen.
#> --- Gordon Hempton, Earth is a Solar Powered Jukebox
bioacousticsquote(pattern="Seuketat")
#> 
#> Seuketat is the Eskimo word for ear-of-the-animal. This is what we must
#> become if we want to truly listen.
#> --- Gordon Hempton, Earth is a Solar Powered Jukebox
bioacousticsquote(source="Hempton")
#> 
#> How can we listen for something that we have not yet heard? We can't.
#> When we listen in this way, selectively, we are actually practicing
#> controlled impairment.
#> --- Gordon Hempton, Earth is a Solar Powered Jukebox
bioacousticsquote(tag="poetry")
#> 
#> Then the little Hiawatha, Learned of every bird its language, Learned
#> their names and all their secrets, How they built their nests in
#> summer, Where they hid themselves in winter, Talked with them whene'er
#> he met them, Called them 'Hiawatha's Chickens.'
#> --- Henry Wadsworth Longfellow, Haiwatha's Childhood
print.data.frame(bioacousticsquote(9)) # All information
#>   qid
#> 9   9
#>                                                                                                         text
#> 9 Sound, because it is so intimate, immediate, and physical, is probably the most influential of our senses.
#>          source             cite  url      tags  tex
#> 9 Bernie Krause Wild Soundscapes <NA> listening <NA>